+++ code/djangoApp/frontend/src/componenets/App.js
@@ -33,6 +33,7 @@ import Tournament from "./Tournament/Tournament";
 import CreateTournament from "./Tournament/CreateTournament";
 import { Container } from "@mui/material";
 import { Box, display } from "@mui/system";
+import Orders from "./Orders/Orders";
 // import BottomAppBar from './BottomAppBar';
 function App() {
     const playerPath = ["Game", "Drinks", "Leaderboards", "Account"]
@@ -68,7 +69,7 @@ function App() {
                                             <GamePage />
                                         }
                                         {role == "drinkMiester" &&
-                                            <Menu />
+                                            <Orders />
                                         }
                                         {role == "manager" &&
                                             <ManageTournamentsPage />
@@ -96,7 +97,7 @@ function App() {
                                             <AccountPage />
                                         </Route>
                                         <Route path="/drink-orders">
-                                            <Menu />
+                                            <Orders />
                                         </Route>
                                         <Route path="/manage-tournaments">
                                             <ManageTournamentsPage />
@@ -123,7 +124,7 @@ function App() {
                                         </Route>
 
                                         <Route path="/manage-sponsorship">
-                                            <ManageSponsorshipPage></ManageSponsorshipPage>
+                                            <ManageSponsorshipPage />
                                         </Route>
                                     </Box>
                                 </Switch>
