+++ code/djangoApp/frontend/src/componenets/Menu/Menu.css
@@ -1,9 +1,3 @@
-.transaction-container{
-    background-color: #ccc;
-    border-radius: 5px;
-    margin-bottom: 5px;
-}
-
 /* Drawer Content */
 .edit-menu-item {
     align-items: center;
