+++ code/djangoApp/frontend/src/componenets/Menu/Menu.js
@@ -1,6 +1,5 @@
 import { useSelector, useDispatch } from "react-redux";
 import React, { useEffect, useMemo, useState } from "react";
-import { getAllActiveTransactions } from "../../services/services";
 import { Drawer } from "@material-ui/core";
 import Icon from '@mui/material/Icon';
 import './Menu.css'
@@ -11,19 +10,9 @@ const Menu = () => {
     const role = useSelector((state) => state.user.role);
     const [toggleAll, setToggleAll] = useState(false);
     const [toggleDrawer, setToggleDrawer] = useState(false);
-    const [transactions, setTransactions] = useState([]);
     const [editable, setEditable] = useState(false);
     const [quantity, setQuantity] = useState(0);
-    const dispatch = useDispatch();
-    useEffect(() => {
-        if (role === 'drinkMiester') {
-            getAllActiveTransactions().then((data) => {
-                setTransactions(data);
-            });
-        }
-    }, []);
-
-
+    
     const handleIconClick = (itemId, edit = false) => {
         setToggleDrawer(true);
         setEditable(edit)
@@ -32,17 +21,15 @@ const Menu = () => {
         // Add item to purchase history
         setToggleDrawer(false);
     }
-    const handleCompleteClick = () => {
-        // Complete transaction
-    }
-    const handleDeleteClick = () => {
-        // Delete transaction
-    }
+    
     const handleAddMenuItemClick = () => {
         // Add menu item
     }
     const handleEditMenuItem = () => {
         // Add menu
+    }
+    const handleDeleteMenuItem = () => {
+        // Delete menu item
     }
     const item = { id: 1 }
     return (
@@ -70,25 +57,8 @@ const Menu = () => {
                     </div>
 
                 </div>
-            }{role === "drinkMiester" &&
-                <div className="drink-miester-view">
-                    {transactions.map((item) => {
-                        return (
-                            <div className="transaction-container" key={item.transactionid}>
-                                <div className="item-name">{item.transactionbuyer}</div>
-                                <div className="item-price">{item.transactiondrinkmeister}</div>
-                                <div className="item-quantity">{item.transactiondate}</div>
-                                <div className="item-quantity">${item.transactionprice}</div>
-                                <div className="item-quantity">{item.transactionactiveflag}</div>
-                                <button className="complete-button" onClick={() => { handleCompleteClick(item.transactionid) }}>Complete</button>
-                                <button className="delete-button" onClick={() => { handleDeleteClick(item.transactionid) }}>Delete</button>
-                            </div>
-                        )
-                    })}
-                    {/* Get all current drinks and display them */}
-                    {/* Button to fulfill and delete */}
-                </div>
-            }{(role === "owner" || role === "manager") &&
+            }
+            {(role === "owner" || role === "manager") &&
                 <div className="manager-view">
                     {/* Edit drinks */}
                     <div className="menu-items">
@@ -106,6 +76,7 @@ const Menu = () => {
                         </div>
                         <div>
                         <button className="add-button" onClick={() => { handleAddMenuItemClick() }}>Add</button>
+                        <button className="delete-button" onClick={() => { handleDeleteMenuItem() }}>Edit</button>
                         </div>
                     </div>
                 </div>
