+++ code/djangoApp/frontend/src/componenets/Orders/Orders.css
@@ -1,0 +1,24 @@
+.transaction-container{
+    background-color: #ccc;
+    border-radius: 5px;
+    margin-bottom: 5px;
+}
+.complete-button {
+    color: #fff;
+    margin: 10px;
+    background-color: #3f51b5;
+    border-radius: 5px;
+    border: none;
+    font-size: 15px;
+    padding: 5px
+}
+
+.delete-button{
+    color: #fff;
+    margin: 10px;
+    background-color: #bd1717;
+    border-radius: 5px;
+    border: none;
+    font-size: 15px;
+    padding: 5px
+}
